fix: use `Long` instead of `Int` for date value

+++ app/src/main/java/xyz/sevive/arcaeaoffline/core/database/entities/ScoreCalculated.kt
@@ -33,7 +33,7 @@ data class ScoreCalculated(
     @ColumnInfo(name = "shiny_pure") val shinyPure: Int?,
     val far: Int?,
     val lost: Int?,
-    val date: Int?,
+    val date: Long?,
     @ColumnInfo(name = "max_recall") val maxRecall: Int?,
     val modifier: Int?,
     @ColumnInfo(name = "clear_type") val clearType: Int?,
